Move compare.go to encoding module
