+++ code.go
@@ -4,6 +4,7 @@ import (
 	"time"
 
 	"github.com/google/orderedcode"
+	"github.com/ostafen/clover/encoding"
 	"github.com/ostafen/clover/util"
 )
 
@@ -26,7 +27,7 @@ func encodePrimitive(buf []byte, value interface{}, includeType bool) ([]byte, e
 
 	actualVal := getEncodeValue(value)
 	if includeType {
-		typeId := uint64(typesMap[getTypeName(value)])
+		typeId := uint64(encoding.TypeId(value))
 		buf, err = orderedcode.Append(buf, typeId)
 		if err != nil {
 			return nil, err
